consolidated error/warning structures into constants.js
