+++ test/test.js
@@ -3,7 +3,7 @@ var xlsx = require("node-xlsx");
 // var cytoscape = require("cytoscape");
 var spreadsheetController = require(__dirname + "/../server/controllers" + "/spreadsheet-controller")();
 
-var parseWorkbookSheet = require(__dirname + "/../server/controllers" + "/workbook-sheet-parser");
+var parseNetworkSheet = require(__dirname + "/../server/controllers" + "/network-sheet-parser");
 
 var parseExpressionSheet = require(__dirname + "/../server/controllers" + "/expression-sheet-parser");
 
@@ -13,13 +13,13 @@ var parseAdditionalSheet = require(__dirname + "/../server/controllers" + "/addi
 
 var noErrors = function (input) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     assert.equal(0, workbook.errors.length);
 };
 
 var duplicateGeneError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
 
     assert.equal(frequency, workbook.errors.length);
 
@@ -39,7 +39,7 @@ var duplicateGeneError = function (input, frequency) {
 
 var invalidGeneLengthError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
 
     assert.equal(frequency, workbook.errors.length);
 
@@ -53,7 +53,7 @@ var invalidGeneLengthError = function (input, frequency) {
 
 var corruptGeneError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
 
     assert.equal(frequency, workbook.errors.length);
 
@@ -67,7 +67,7 @@ var corruptGeneError = function (input, frequency) {
 
 var unknownError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
 
     assert.equal(frequency, workbook.errors.length);
 
@@ -81,7 +81,7 @@ var unknownError = function (input, frequency) {
 
 var missingValueError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
 
     assert.equal(frequency, workbook.errors.length);
 
@@ -95,7 +95,7 @@ var missingValueError = function (input, frequency) {
 
 var missingNetworkError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
 
     assert.equal(frequency, workbook.errors.length);
 
@@ -109,7 +109,7 @@ var missingNetworkError = function (input, frequency) {
 
 var specialCharacterError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
 
     assert.equal(frequency, workbook.errors.length);
 
@@ -123,7 +123,7 @@ var specialCharacterError = function (input, frequency) {
 
 var invalidDataTypeError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
 
     assert.equal(frequency, workbook.errors.length);
 
@@ -137,7 +137,7 @@ var invalidDataTypeError = function (input, frequency) {
 
 var workbookSizeError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
 
     assert.equal(frequency, workbook.errors.length);
 
@@ -151,7 +151,7 @@ var workbookSizeError = function (input, frequency) {
 
 var checkForGene = function (test, frequency, input) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
 
     assert.equal(frequency, workbook.genes.filter(function (gene) {
         return gene.name === test;
@@ -160,7 +160,7 @@ var checkForGene = function (test, frequency, input) {
 
 var warningsCountError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var warningsCountErrorArray = workbook.errors.filter(function (x) {
         return x.errorCode === "WARNINGS_OVERLOAD";
     });
@@ -170,7 +170,7 @@ var warningsCountError = function (input, frequency) {
 
 var errorsCountError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var errorsCountErrorArray = workbook.errors.filter(function (x) {
         return x.errorCode === "ERRORS_OVERLOAD";
     });
@@ -180,7 +180,7 @@ var errorsCountError = function (input, frequency) {
 
 var emptyRowError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
 
     assert.equal(frequency, workbook.errors.length);
 
@@ -323,7 +323,7 @@ var nonNumericalTimePointError = function (input, frequency) {
 
 var emptyRowDataError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var emptyRowDataCount = workbook.errors.filter(function (x) {
         return x.errorCode === "EMPTY_ROW_DATA";
     });
@@ -333,7 +333,7 @@ var emptyRowDataError = function (input, frequency) {
 
 var emptyMatrixDataError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var emptyMatrixDataCount = workbook.errors.filter(function (x) {
         return x.errorCode === "EMPTY_MATRIX_DATA";
     });
@@ -343,7 +343,7 @@ var emptyMatrixDataError = function (input, frequency) {
 
 var emptyColumnDataError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var emptyRowDataCount = workbook.errors.filter(function (x) {
         return x.errorCode === "EMPTY_COLUMN_DATA";
     });
@@ -353,7 +353,7 @@ var emptyColumnDataError = function (input, frequency) {
 
 var emptyColumnError = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var emptyRowDataCount = workbook.errors.filter(function (x) {
         return x.errorCode === "EMPTY_COLUMN";
     });
@@ -367,14 +367,14 @@ var emptyColumnError = function (input, frequency) {
 
 var noWarnings = function (input) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
 
     assert.equal(0, workbook.warnings.length);
 };
 
 var missingSourceWarning = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var missingSourceCount = workbook.warnings.filter(function (x) {
         return x.warningCode === "MISSING_SOURCE";
     });
@@ -384,7 +384,7 @@ var missingSourceWarning = function (input, frequency) {
 
 var invalidMatrixDataWarning = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var invalidDataCount = workbook.warnings.filter(function (x) {
         return x.warningCode === "INVALID_DATA";
     });
@@ -394,7 +394,7 @@ var invalidMatrixDataWarning = function (input, frequency) {
 
 var missingTargetWarning = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var missingTargetCount = workbook.warnings.filter(function (x) {
         return x.warningCode === "MISSING_TARGET";
     });
@@ -404,7 +404,7 @@ var missingTargetWarning = function (input, frequency) {
 
 var randomDataWarning = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var randomDataCount = workbook.warnings.filter(function (x) {
         return x.warningCode === "RANDOM_DATA";
     });
@@ -414,7 +414,7 @@ var randomDataWarning = function (input, frequency) {
 
 var emptyRowWarning = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var emptyRowCount = workbook.warnings.filter(function (x) {
         return x.warningCode === "EMPTY_ROW";
     });
@@ -424,7 +424,7 @@ var emptyRowWarning = function (input, frequency) {
 
 var invalidNetworkSizeWarning = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var invalidworkbookSizeCount = workbook.warnings.filter(function (x) {
         return x.warningCode === "INVALID_NETWORK_SIZE";
     });
@@ -465,7 +465,7 @@ var incorrectlyNamedExpressionSheetWarning = function (input, frequency) {
 
 var incorrectlyNamedSheetWarning = function (input, frequency) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var incorrectlyNamedSheetCount = workbook.warnings.filter(function (x) {
         return x.warningCode === "INCORRECTLY_NAMED_SHEET";
     });
@@ -477,7 +477,7 @@ var incorrectlyNamedSheetWarning = function (input, frequency) {
 /*
 var shortestPath = function (input, directed, source, target, length) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var cytoscapeElements = spreadsheetController.grnSightToCytoscape(workbook);
 
     var cy = cytoscape({
@@ -491,7 +491,7 @@ var shortestPath = function (input, directed, source, target, length) {
 
 var betweennessCentrality = function (input, directed, node, centrality) {
     var sheet = xlsx.parse(input);
-    var workbook = parseWorkbookSheet(sheet);
+    var workbook = parseNetworkSheet(sheet);
     var cytoscapeElements = spreadsheetController.grnSightToCytoscape(workbook);
 
     var cy = cytoscape({
